Much better Bolts::Bag API
I tried to document how to use the other and aside from the magic it did in
Bolts::Meta::Locator, I couldn't figure out how to write a synopsis. I took that
as a design error and repaired the problem. Gone is the singular create methods.
Instead, you use the class to start the bag, which returns a partially
constructed metaclass you can fill in and then finish and construct yourself.