While adding docs to injectors, found a couple of issues
The terms dependencies, parameters, etc. are too overloaded. There needs to be a
better separation of terminology to avoid confusion. Also, there's no way to use
the Bolts sugar to inject via setter or store in an array or hash. To fix this,
I need a better term than dependency. First, renaming many uses of dependencies
to parameters.