More nomenclature fixes, this time to inferrence
The words "parameters" and "dependencies" are wrong. I changed "parameters" to
"options" to help make it clearer that this means parameters passed to
acquire(). Similarly, I renamed "dependencies" to "acquisition" to note that
these will be automatically acquired. I considered "automatic" but thought that
might be too vague as there could be other kinds of automatic inferrence in the
future. At least, that's how I justify the choice I made in about 2 minutes
time.